+++ ec2_Yukon_GDPS/setup_gem_forecast_mod.sh
@@ -47,6 +47,7 @@ do
    # Obtain other mesh files that are unchanged from the previous gem forecast run
    cp ${input_file_path_gem}/MESH_drainage_database.r2c ${output_file_path_gem}
    [[ -f ${input_file_path_gem}/MESH_input_reservoir.txt ]] && cp ${input_file_path_gem}/MESH_input_reservoir.txt ${output_file_path_gem}
+   [[ -f ${input_file_path_gem}/MESH_input_reservoir.tb0 ]] && cp ${input_file_path_gem}/MESH_input_reservoir.tb0 ${output_file_path_gem}
    cp ${input_file_path_gem}/MESH_input_soil_levels.txt ${output_file_path_gem}
    cp ${input_file_path_gem}/MESH_parameters_hydrology.ini ${output_file_path_gem}
    [[ -f ${input_file_path_gem}/MESH_parameters.r2c ]] && cp ${input_file_path_gem}/MESH_parameters.r2c ${output_file_path_gem}
@@ -59,7 +60,8 @@ do
    cp ${input_file_path_gem}/MESH_input_run_options_180.ini ${output_file_path_gem}
    cp ${input_file_path_gem}/MESH_parameters_CLASS.ini ${output_file_path_gem}
    # The date and streamflow values in this file needs changing
-   cp ${input_file_path_gem}/MESH_input_streamflow.txt ${output_file_path_gem} 
+   [[ -f ${input_file_path_gem}/MESH_input_streamflow.txt ]] && cp ${input_file_path_gem}/MESH_input_streamflow.txt ${output_file_path_gem} 
+   [[ -f ${input_file_path_gem}/MESH_input_streamflow.tb0 ]] && cp ${input_file_path_gem}/MESH_input_streamflow.tb0 ${output_file_path_gem} 
 
    # Obtain mesh state files from the previous capa run
    if [ -f ${input_file_path_capa}/int_statVariables.seq.rte ]; then
